converted old api rule to react query

+++ client/src/dashboard/card/task.component.tsx
@@ -3,6 +3,7 @@ import { TaskState } from "../task.type";
 import TaskForm from "./task.form";
 
 import axios, { SERVER_URL } from "../../api/http"
+import { useMutation } from "react-query";
 
 export default function Card(props: TaskState) {
     const [createTaskFlag, setCreateTaskFlag] = useState<boolean>(false)
@@ -12,22 +13,48 @@ export default function Card(props: TaskState) {
     const createFlagOff = () => {
         setCreateTaskFlag(false)
     }
+
+    const afterSubmit = () => {
+        if (props.onSubmit) props.onSubmit(true)
+    }
+
     const onSubmit = () => {
-        props.onSubmit(true)
+        afterSubmit()
         createFlagOff()
     }
-    const updateStatus = async () => {
-        await axios.put(`${SERVER_URL}/tasks/${props.id}`,
-            {
-                status: 'COMPLETED'
+    
+    const updateCard = async (isDelete: boolean) => {
+        try {
+            if (isDelete) {
+                const response = await axios.delete(`${SERVER_URL}/tasks/${props.id}`)
+                return response.data
+            } else {
+                const response = await axios.put(`${SERVER_URL}/tasks/${props.id}`,
+                    {
+                        status: 'COMPLETED'
+                    }
+                )
+                return response.data
             }
-        );
-        props.onSubmit(true)
+        } catch(err) {
+            throw new Error('Something went wrong')
+        }
     }
-    const deleteTask = async () => {
-        await axios.delete(`${SERVER_URL}/tasks/${props.id}`)
-        props.onSubmit(true)
+
+
+    const { mutate } = useMutation(updateCard, {
+        onSuccess: () => {
+            afterSubmit()
+        }
+    })
+
+    const updateStatus = () => {
+        mutate(false)
     }
+    const deleteTask = () => {
+        mutate(true)
+    }
+
     return (
         <>
             {
